+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/request/ApplicationStatusReqDto.java
@@ -3,7 +3,7 @@ package team.themoment.hellogsmv3.domain.application.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import team.themoment.hellogsmv3.domain.oneseo.entity.type.EvaluationStatus;
-import team.themoment.hellogsmv3.domain.application.type.Major;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.Major;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 
 import java.math.BigDecimal;
