+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/response/AdmissionStatusResDto.java
@@ -2,7 +2,7 @@ package team.themoment.hellogsmv3.domain.application.dto.response;
 
 import lombok.Builder;
 import team.themoment.hellogsmv3.domain.oneseo.entity.type.EvaluationStatus;
-import team.themoment.hellogsmv3.domain.application.type.Major;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.Major;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 
 import java.math.BigDecimal;
