--- src/main/java/team/themoment/hellogsmv3/domain/application/type/Major.java
@@ -1,7 +1,0 @@
-package team.themoment.hellogsmv3.domain.application.type;
-
-public enum Major {
-    AI,
-    IOT,
-    SW
-}
