+++ src/main/java/team/themoment/hellogsmv3/domain/oneseo/entity/EntranceTestResult.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import team.themoment.hellogsmv3.domain.oneseo.entity.type.EvaluationStatus;
-import team.themoment.hellogsmv3.domain.application.type.Major;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.Major;
 
 import java.math.BigDecimal;
