+++ src/main/java/team/themoment/hellogsmv3/domain/oneseo/entity/type/Major.java
@@ -1,0 +1,7 @@
+package team.themoment.hellogsmv3.domain.oneseo.entity.type;
+
+public enum Major {
+    AI,
+    IOT,
+    SW
+}
